@@ -3,23 +3,38 @@ package com.vaani.leetcode.dp;
 import java.util.Arrays;
 
 /**
- * 4/3/2017. Given a string s, partition s such that every substring of the
- * partition is a palindrome.
- *
- * <p>Return the minimum cuts needed for a palindrome partitioning of s.
- *
- * <p>For example, given s = "aab", Return 1 since the palindrome partitioning ["aa","b"] could be
- * produced using 1 cut.
+ * https://leetcode.com/problems/palindrome-partitioning-ii/
+ * 132. Palindrome Partitioning II
+ * Hard
+ * <p>
+ * Given a string s, partition s such that every substring of the partition is a palindrome.
+ * <p>
+ * Return the minimum cuts needed for a palindrome partitioning of s.
+ * <p>
+ * <p>
+ * <p>
+ * Example 1:
+ * <p>
+ * Input: s = "aab"
+ * Output: 1
+ * Explanation: The palindrome partitioning ["aa","b"] could be produced using 1 cut.
+ * Example 2:
+ * <p>
+ * Input: s = "a"
+ * Output: 0
+ * Example 3:
+ * <p>
+ * Input: s = "ab"
+ * Output: 1
+ * <p>
+ * <p>
+ * Constraints:
+ * <p>
+ * 1 <= s.length <= 2000
+ * s consists of lower-case English letters only.
  */
 public class PalindromePartitioningII {
-    private int A[];
-    private boolean[][] paliandrome;
 
-    /**
-     * Main method
-     *
-     * @param args
-     */
     public static void main(String[] args) {
         System.out.println(
                 new PalindromePartitioningII()
@@ -27,28 +42,37 @@ public class PalindromePartitioningII {
                                 "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaabbaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"));
     }
 
+    private int[] dp;
+    private boolean[][] palindrome;
+
     public int minCut(String s) {
-        if (s == null || s.isEmpty()) return 0;
-        A = new int[s.length()];
-        Arrays.fill(A, Integer.MAX_VALUE);
+        if (s == null || s.isEmpty()) {
+            return 0;
+        }
+        dp = new int[s.length()];
+        Arrays.fill(dp, Integer.MAX_VALUE);
         char[] charArr = s.toCharArray();
-        paliandrome = new boolean[charArr.length][charArr.length];
+        palindrome = new boolean[charArr.length][charArr.length];
         return doNext(charArr, 0) - 1;
     }
 
     private int doNext(char[] s, int p) {
-        if (p >= s.length) return 0;
-        if (A[p] < Integer.MAX_VALUE) return A[p];
+        if (p >= s.length) {
+            return 0;
+        }
+        if (dp[p] < Integer.MAX_VALUE) {
+            return dp[p];
+        }
         for (int i = p, l = s.length; i < l; i++) {
             if (p + 1 <= i - 1) {
-                paliandrome[p][i] = (paliandrome[p + 1][i - 1] && (s[p] == s[i]));
+                palindrome[p][i] = (palindrome[p + 1][i - 1] && (s[p] == s[i]));
             } else {
-                paliandrome[p][i] = (i == p) || (s[i] == s[p]);
+                palindrome[p][i] = (i == p) || (s[i] == s[p]);
             }
-            if (paliandrome[p][i]) {
-                A[p] = Math.min(doNext(s, i + 1) + 1, A[p]);
+            if (palindrome[p][i]) {
+                dp[p] = Math.min(doNext(s, i + 1) + 1, dp[p]);
             }
         }
-        return A[p];
+        return dp[p];
     }
 }
