@@ -1,45 +1,59 @@
 package com.vaani.leetcode.heap;
 
+import java.util.Comparator;
 import java.util.PriorityQueue;
 
 /**
- * 26/07/2018. There are N children standing in a line. Each child
- * is assigned a rating value.
- *
- * <p>You are giving candies to these children subjected to the following requirements:
- *
- * <p>Each child must have at least one candy. Children with a higher rating get more candies than
- * their neighbors. What is the minimum candies you must give?
- *
- * <p>Example 1:
- *
- * <p>Input: [1,0,2] Output: 5 Explanation: You can allocate to the first, second and third child
- * with 2, 1, 2 candies respectively. Example 2:
- *
- * <p>Input: [1,2,2] Output: 4 Explanation: You can allocate to the first, second and third child
- * with 1, 2, 1 candies respectively. The third child gets 1 candy because it satisfies the above
- * two conditions.
- *
- * <p>Solution: O(N log N): Store the indexes in a com.vaani.leetcode.heap, iterate through the com.vaani.leetcode.heap one by one and
- * assign candies one greater than its neighbours. Take care of vertex cases.
+ * https://leetcode.com/problems/candy/
+ * 135. Candy
+ * Hard
+ * <p>
+ * There are n children standing in a line. Each child is assigned a rating value given in the integer array ratings.
+ * <p>
+ * You are giving candies to these children subjected to the following requirements:
+ * <p>
+ * Each child must have at least one candy.
+ * Children with a higher rating get more candies than their neighbors.
+ * Return the minimum number of candies you need to have to distribute the candies to the children.
+ * <p>
+ * <p>
+ * <p>
+ * Example 1:
+ * <p>
+ * Input: ratings = [1,0,2]
+ * Output: 5
+ * Explanation: You can allocate to the first, second and third child with 2, 1, 2 candies respectively.
+ * Example 2:
+ * <p>
+ * Input: ratings = [1,2,2]
+ * Output: 4
+ * Explanation: You can allocate to the first, second and third child with 1, 2, 1 candies respectively.
+ * The third child gets 1 candy because it satisfies the above two conditions.
+ * <p>
+ * <p>
+ * Constraints:
+ * <p>
+ * n == ratings.length
+ * 1 <= n <= 2 * 10^4
+ * 0 <= ratings[i] <= 2 * 10^4
  */
 public class Candy {
-
-    /**
-     * Main method
-     *
-     * @param args
-     * @throws Exception
-     */
     public static void main(String[] args) throws Exception {
         int[] ratings = {29, 51, 87, 87, 72, 12};
         System.out.println(new Candy().candy(ratings));
     }
 
+    /*
+     *
+     * <p>Solution: O(N log N): Store the indexes in a com.vaani.leetcode.heap, iterate through the com.vaani.leetcode.heap one by one and
+     * assign candies one greater than its neighbours. Take care of vertex cases.
+     */
     public int candy(int[] ratings) {
-        if (ratings.length == 1) return 1;
+        if (ratings.length == 1) {
+            return 1;
+        }
         PriorityQueue<Integer> pq =
-                new PriorityQueue<>((o1, o2) -> Integer.compare(ratings[o1], ratings[o2]));
+                new PriorityQueue<>(Comparator.comparingInt(o -> ratings[o]));
         for (int i = 0; i < ratings.length; i++) {
             pq.offer(i);
         }
@@ -82,10 +96,10 @@ public class Candy {
                 }
             }
         }
-        int result = 0;
+        int ans = 0;
         for (int c : count) {
-            result += c;
+            ans += c;
         }
-        return result;
+        return ans;
     }
 }
