@@ -1,6 +1,6 @@
 package com.vaani.leetcode.tree;
 
-import com.vaani.dsa.ds.core.tree.binarytree.simple.BinaryTreeNode;
+import com.vaani.dsa.ds.core.tree.binarytree.simple.TreeNode;
 
 /**
  * Given a binary search tree and a node in it, find
@@ -16,26 +16,26 @@ import com.vaani.dsa.ds.core.tree.binarytree.simple.BinaryTreeNode;
 public class InorderSuccessorInBST {
 
     public static void main(String[] args) throws Exception {
-        BinaryTreeNode root = new BinaryTreeNode(10);
-        root.left = new BinaryTreeNode(5);
-        root.left.left = new BinaryTreeNode(3);
-        root.left.right = new BinaryTreeNode(6);
-        root.right = new BinaryTreeNode(15);
-        root.right.left = new BinaryTreeNode(13);
-        root.right.left.left = new BinaryTreeNode(12);
-        root.right.left.right = new BinaryTreeNode(14);
-        root.right.right = new BinaryTreeNode(17);
-        BinaryTreeNode ans = new InorderSuccessorInBST().inorderSuccessor(root, root.right.left.right);
+        TreeNode root = new TreeNode(10);
+        root.left = new TreeNode(5);
+        root.left.left = new TreeNode(3);
+        root.left.right = new TreeNode(6);
+        root.right = new TreeNode(15);
+        root.right.left = new TreeNode(13);
+        root.right.left.left = new TreeNode(12);
+        root.right.left.right = new TreeNode(14);
+        root.right.right = new TreeNode(17);
+        TreeNode ans = new InorderSuccessorInBST().inorderSuccessor(root, root.right.left.right);
         if (ans != null) System.out.println(ans.val);
         else System.out.println(ans);
     }
 
-    public BinaryTreeNode inorderSuccessor(BinaryTreeNode root, BinaryTreeNode p) {
+    public TreeNode inorderSuccessor(TreeNode root, TreeNode p) {
         return inOrder(root, p, null);
     }
 
 
-    private BinaryTreeNode inOrder(BinaryTreeNode curr, BinaryTreeNode target, BinaryTreeNode successor) {
+    private TreeNode inOrder(TreeNode curr, TreeNode target, TreeNode successor) {
         if (curr == null) return successor;
         if (curr.val <= target.val) return inOrder(curr.right, target, successor);
         return inOrder(curr.left, target, curr); // make the current node as successor
